requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client for callbacks
httpx==0.26.0

# Google Gemini API
google-generativeai>=0.7.0

# Python Dotenv for environment variables
python-dotenv==1.0.0

# Additional utilities
python-multipart==0.0.6
